trying to make the sound in qemu work.

+++ Kernel/syscalls.c
@@ -69,7 +69,7 @@ void _9_set_font_size(uint64_t size) {
 }
 
 void _10_beep(uint64_t frequency){
-    beep(frequency);
+    beep((int)frequency);
 }
 
 void _11_get_ticks(uint64_t  delta) {
